re-open the mixer at a different freq to avoid resampling
resampling must be avoided when there are too many tracks to resample.
Re-opening the mixer gives exactly the same sound, but with no resampling,
which saves a lot of cpu power.
The only drawback : the sounds played directly (OneSound) play at a bad
frequency, but when it's 48Khz instead of 44.1, it's almost impossible to
hear, it can be heard when training though...