resampling of short sounds only with scikits.samplerate
scipy is way too slow for that, the fact that scikits can handle directly
stereo samples is probably a big part of why...
In case scikits.samplerate is not installed, resampling is disabled.
Also the sounds are not kept in ram, even when resampling, it eats way too
much ram.