Adjustment for old primitive songs
it was breaking recent phase shift ones using unrecognized tracks.
Solution for now : listing all the unrecognized tracks in the old songs.
It's possible to miss a few by doing so, but it's the safest way I have
found for now. All the sectoid + muldjord are done, which are parts of the
debian distribution for fof, so it's already a big part of it.